Split data: palette
