+++ extension/devtools/app/lib/main.dart
@@ -27,6 +27,7 @@ part 'filters.dart';
 part 'status_styles.dart';
 part 'color_maps.dart';
 part 'effects_timeline_constants.dart';
+part 'palette.dart';
 part 'widgets/adaptive_wrap.dart';
 part 'widgets/action_buttons.dart';
 part 'widgets/action_pill.dart';
@@ -207,15 +208,4 @@ class OrefTheme {
       cardTheme: const CardThemeData(color: Colors.transparent, elevation: 0),
     );
   }
-}
-
-class OrefPalette {
-  static const Color teal = Color(0xFF22E3C4);
-  static const Color tealDark = Color(0xFF14B6A1);
-  static const Color indigo = Color(0xFF6C5CFF);
-  static const Color coral = Color(0xFFFF8C6B);
-  static const Color lime = Color(0xFFB5FF6D);
-  static const Color pink = Color(0xFFFF71C6);
-  static const Color deepBlue = Color(0xFF0C141C);
-  static const Color lightBlue = Color(0xFFE7F3FF);
 }
