+++ extension/devtools/app/lib/palette.dart
@@ -1,0 +1,12 @@
+part of 'main.dart';
+
+class OrefPalette {
+  static const Color teal = Color(0xFF22E3C4);
+  static const Color tealDark = Color(0xFF14B6A1);
+  static const Color indigo = Color(0xFF6C5CFF);
+  static const Color coral = Color(0xFFFF8C6B);
+  static const Color lime = Color(0xFFB5FF6D);
+  static const Color pink = Color(0xFFFF71C6);
+  static const Color deepBlue = Color(0xFF0C141C);
+  static const Color lightBlue = Color(0xFFE7F3FF);
+}
